chore(tests): fix `CourseOutline` tests
The previous test was looking for an element with the role `alert`. It was passing because it found the  `altert` role on the `ToastContainer`, which was removed in https://github.com/openedx/paragon/pull/3434.

This means the test wasn't actually verifying that the application "displays an alert"

It now just looks for an element with the test id `redux-provider` which serves the same purpose of ensuring the application has reached a state where the `loadingStatus` will be properly set to DENIED.