refactor: update SearchFilterWidget to use Dropdown component
Replaced the ModalPopup with a Dropdown for better user experience. The filtering dropdowns (publish status/ component type/ tags) all close out if a user clicks anywhere outside of the dropdown menus. In this implementation, however, clicking outside of the opened menu only triggers closing the menu. As in: if a filtering dropdown is open, and decide you want to open a different filtering option, a user has to click twice: once to close the initially opened dropdown, and a second time to open the second filtering option.