fix: enable publish button after component edit
This was fixed in master in commit https://github.com/openedx/frontend-app-authoring/commit/ad62519af75ffcd797dc8356fff28badfda72f7a but at that time, the changes of current PR also contributed to the fix. We backported the fix to release/teak but that didn't work because of additional changes that this commit adds. We avoided the backport of the original PR due to additional refactoring.